[INDDEX] Add 'Master Data File' report
