+++ custom/inddex/ucr/data_providers/gaps_summary_data.py
@@ -3,564 +3,6 @@ from memoized import memoized
 from custom.inddex.ucr.data_providers.mixins import GapsReportSummaryDataMixin
 
 
-class GapsSummaryMasterOutputData(GapsReportSummaryDataMixin):
-    title = 'Master Output'
-    slug = 'master_output'
-    headers_in_order = [
-        'unique_respondent_id', 'recall_case_id', 'opened_date', 'opened_by_username', 'owner_name', 'recall_date',
-        'recall_status', 'gender', 'age_years', 'age_months', 'age_range', 'pregnant', 'breastfeeding',
-        'supplements', 'urban_rural', 'food_code', 'food_name', 'recipe_name', 'caseid', 'ingr_recipe_case_id',
-        'reference_food_code', 'food_type', 'include_in_analysis', 'food_status', 'eating_time', 'time_block',
-        'fao_who_gift_food_group_code', 'fao_who_gift_food_group_description', 'user_food_group',
-        'already_reported_food', 'already_reported_food_caseid', 'recipe_already_reported',
-        'recipe_already_reported_caseid', 'recipe_already_reported_name', 'is_ingredient', 'ingr_recipe_case_id',
-        'ingr_recipe_code', 'ingr_fraction', 'ingr_recipe_total_grams_consumed', 'short_name', 'food_base_term',
-        'tag_1', 'tag_1_other', 'tag_2', 'tag_2_other', 'tag_3', 'tag_3_other', 'tag_4', 'tag_4_other', 'tag_5',
-        'tag_5_other', 'tag_6', 'tag_6_other', 'tag_7', 'tag_7_other', 'tag_8', 'tag_8_other', 'tag_9',
-        'tag_9_other', 'tag_10', 'tag_10_other', 'conv_method_code', 'conv_method_desc', 'conv_option_code',
-        'conv_option_desc', 'conv_size', 'conv_units', 'quantity', 'nsr_conv_method_code_post_cooking',
-        'nsr_conv_method_desc_post_cooking', 'nsr_conv_option_code_post_cooking',
-        'nsr_conv_option_desc_post_cooking', 'nsr_conv_size_post_cooking', 'nsr_consumed_cooked_fraction',
-        'recipe_num_ingredients', 'conv_factor_food_code', 'conv_factor_base_term_food_code', 'conv_factor_used',
-        'fct_food_code_exists', 'fct_base_term_food_code_exists', 'fct_reference_food_code_exists',
-        'fct_data_used', 'total_grams', 'energy_per_100g', 'energy_kcal', 'water_per_100g', 'water_g',
-        'protein_per_100g', 'protein_g', 'conv_factor_gap_code', 'conv factor_gap_desc', 'fct_gap_code',
-        'fct_gap_desc'
-    ]
-
-    def __init__(self, config):
-        self.config = config
-
-    @property
-    def filter_values(self):
-        return self.filters_config
-
-    @property
-    @memoized
-    def rows(self):
-        return [
-            ['555_1', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b13', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '13017', 'Mắm,Nước mắm cá', '',
-             '548ea376-28a1-43e2-8df2-b26ca445d23f', '', '1301199', 'food_item', 'yes', 'fourth_pass', '07:46',
-             '6 am - 8:59 am', '13', 'Spices and condiments (13)', '', '', '', '', '', '', '', '', '', '', '',
-             'nước mắm chấm', 'Mắm', 'Nước mắm cá', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', '50', 'Proxy weight (g) - water', '', '', '5', 'grams', '1', '', '', '', '', '',
-             '', '',
-             '1', '1', 'food_code', 'yes', 'yes', 'yes', 'reference_food_code', '5', '34.89', '1.7445', '51.2',
-             '2.56',
-             '18.5', '0.925', '1', 'conv factor available', '3', 'using fct data from reference food code'],
-            ['666_1', '36271593-cb0d-4a99-ac35-260db7d4b63e', '2019-08-22', 'sue', 'sue', '2019-08-14', 'Complete',
-             'female', '29', '348', '15-49 years', 'no', 'no', 'yes', 'peri-urban', '7170', 'Nước luộc,rau', '',
-             '63069d9d-e46d-4566-b788-b794a2c7f67a', '', '712499', 'food_item', 'yes', 'fourth_pass', '07:46',
-             '6 am - 8:59 am', '1', 'Cereals and their products (1)', '', '', '', '', '', '', '', '', '', '', '',
-             'nước rau cải bắp luộc', 'Nước luộc', 'rau', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW',
-             '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', '50', 'Proxy weight (g) - water', '', '', '299', 'grams', '1', '', '', '', '', '',
-             '',
-             '', '1', '1.0337', 'food_code', 'no', 'yes', 'no', 'base_term_food_code', '299', '0', '0', '5.2',
-             '15.548',
-             '16.1', '48.139', '1', 'conv factor available', '2', 'using fct data from base term food code'],
-            ['666_1', '36271593-cb0d-4a99-ac35-260db7d4b63e', '2019-08-22', 'sue', 'sue', '2019-08-14', 'Complete',
-             'female', '29', '348', '15-49 years', 'no', 'no', 'yes', 'peri-urban', '1004002', 'Cơm,tẻ,nấu', '',
-             'ea7ea242-b196-4cfb-8e65-2bfe66b8e199', '', '100400299', 'food_item', 'yes', 'fourth_pass', '06:30',
-             '6 am - 8:59 am', '3', 'Pulses, seeds and nuts and their products (3)', '', 'yes',
-             'df6c2899-2730-4ff3-b87e-a6b076eb5a0e', '', '', '', '', '', '', '', '', 'cơm', 'Cơm', 'tẻ', '', 'nấu',
-             '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '51', 'Proxy weight (g) - rice', '', '',
-             '172',
-             'grams', '1', '', '', '', '', '', '', '', '0.8588', '0.8588', 'food_code', 'no', 'no', 'no', 'none',
-             '147.7136', '', '', '', '', '', '', '1', 'conv factor available', '8', 'no fct data available'],
-            ['555_2', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b14', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '8058', 'Ruốc,cá', '',
-             '126394a1-a907-45fb-9249-663216062664', '', '805999', 'food_item', 'yes', 'fourth_pass', '19:20',
-             '6 pm - 8:59 pm', '6', 'Fish, shellfish and their products (6)', '', '', '', '', '', '', '', '', '',
-             '',
-             '', 'ruốc cá', 'Ruốc', 'cá', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW',
-             '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             '51',
-             'Proxy weight (g) - rice', '', '', '52', 'grams', '1', '', '', '', '', '', '', '', '', '1.1765',
-             'base_term_food_code', 'yes', 'yes', 'no', 'food_code', '61.178', '311.7', '190.691826', '67.8',
-             '41.478684', '51.7', '31.629026', '2', 'using conversion factor from base term food code', '1',
-             'fct data available'],
-            ['666_1', '36271593-cb0d-4a99-ac35-260db7d4b63e', '2019-08-22', 'sue', 'sue', '2019-08-14', 'Complete',
-             'female', '29', '348', '15-49 years', 'no', 'no', 'yes', 'urban', '', 'Nước luộc,Khác', '',
-             '667c59d9-eebf-4073-be45-9c73950415af', '', '712499', 'food_item', 'yes', 'fourth_pass', '04:05',
-             '3 am - 5:59 am', '4', 'Milk and milk products (4)', '', '', '', '', '', '', '', '', '', '', '',
-             'nước rau luộc', 'Nước luộc', 'Khác', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', '50', 'Proxy weight (g) - water', '', '', '200', 'grams', '1', '', '', '', '', '',
-             '',
-             '', '', '1.0337', 'base_term_food_code', 'no', 'yes', 'no', 'base_term_food_code', '206.74', '0', '0',
-             '5.2', '10.75048', '16.1', '33.28514', '2', 'using conversion factor from base term food code', '2',
-             'using fct data from base term food code'],
-            ['888_1', 'de16bd5b-51bc-4dac-ba81-50e04ac858e5', '2019-08-22', 'mary', 'mary', '2019-08-14',
-             'Complete',
-             'female', '29', '348', '15-49 years', 'yes', 'no', 'yes', 'rural', '', 'Bánh mỳ,que,không ngọt', '',
-             '499f034b-95e3-4d02-909d-af382b1cd46e', '', '101299', 'food_item', 'yes', 'fourth_pass', '09:42',
-             '9 am - 11:59 am', '5', 'Eggs and their products (5)', '', '', '', '', '', '', '', '', '', '', '',
-             'bánh mì trắng', 'Bánh mỳ', 'que', '', 'không ngọt', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', '10', 'Direct weight (g)', '', '', '250', 'grams', '1', '', '', '', '', '', '', '',
-             '',
-             '1', 'base_term_food_code', 'no', 'no', 'no', 'none', '250', '', '', '', '', '', '', '2',
-             'using conversion factor from base term food code', '8', 'no fct data available'],
-            ['8003_1', '832b98ad-303f-458d-a400-1f8453b358c6', '2019-08-22', 'mary', 'mary', '2019-08-14',
-             'Complete',
-             'male', '52', '624', '50-64 years', 'no', 'no', 'no', 'urban', '5017', 'Đu đủ chín', '',
-             '1e1d14d3-a801-4249-ab62-c05f2e92e53b', '', '5017', 'food_item', 'yes', 'fourth_pass', '19:33',
-             '6 pm - 8:59 pm', '10', 'Fruits and their products (10)', '', 'yes',
-             '78e1da3e-506a-407d-b040-fe8b9b10d25d', '', '', '', '', '', '', '', '', 'đu đủ', 'Đu đủ chín',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '10',
-             'Direct weight (g)', '', '', '154', 'grams', '1', '', '', '', '', '', '', '', '', '', 'none', 'yes',
-             'yes',
-             'no', 'food_code', '', '37.20000000000005', '', '67.6', '', '65.9', '', '8',
-             'no conversion factor available', '1', 'fct data available'],
-            ['555_1', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b13', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '8039014', 'Mực khô,nướng', '',
-             'a52b3be8-e445-4705-bfef-73942ae340a2', '', '803999', 'food_item', 'yes', 'fourth_pass', '01:21',
-             '12 pm - 2:59 pm', '6', 'Fish, shellfish and their products (6)', '', '', '', '', '', '', '', '', '',
-             '',
-             '', 'mực khô nướng', 'Mực khô', 'nướng', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', '90', 'Photo substitutes (g)', '1', 'NC06-01', '1', 'grams', '1', '', '', '', '',
-             '',
-             '', '', '59.9155', '80.64', 'food_code', 'yes', 'yes', 'no', 'food_code', '59.9155', '290.9',
-             '174.2941895', '45.4', '27.201637', '78.3', '46.9138365', '1', 'conv factor available', '1',
-             'fct data available'],
-            ['12345_1', 'fbcc5cd9-3d8c-43c8-908d-abb1c0b21089', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'male', '23', '276', '15-49 years', 'yes', 'no', 'no', 'urban', '4116', 'Dưa,cải bẹ', '',
-             '964416e8-6a98-49d4-86df-cd667bad102a', '', '411599', 'food_item', 'yes', 'fourth_pass', '09:19',
-             '9 am - 11:59 am', '7', 'Meat and meat products (7)', '', '', '', '', '', '', '', '', '', '', '',
-             'dưa muối', 'Dưa', 'cải bẹ', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW',
-             '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             '90',
-             'Photo substitutes (g)', '7', 'KC04-07', '1', 'grams', '7', '', '', '', '', '', '', '', '82.08',
-             '152',
-             'food_code', 'no', 'yes', 'no', 'base_term_food_code', '574.56', '24.39999999999997', '140.19264',
-             '72.1',
-             '414.25776', '8.9', '51.13584', '1', 'conv factor available', '2',
-             'using fct data from base term food code'],
-            ['8007_1', '225281cf-ae7e-4d20-9516-0e64c228467e', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'female', '1', '14', '6-59 months', 'no', 'yes', 'yes', 'rural', '5040', 'Ổi', '',
-             'ac63f5a8-4aec-4551-a843-e7f0b5f10515', '', '5040', 'food_item', 'yes', 'fourth_pass', '04:51',
-             '3 am - 5:59 am', '', '', '', '', '', '', '', '', '', '', '', '', '', 'ổi', 'Ổi', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '40', 'Photos (g)', '4',
-             'NC07-03', '1', 'grams', '1', '', '', '', '', '', '', '', '34.2', '34.2', 'food_code', 'no', 'no',
-             'no',
-             'none', '34.2', '', '', '', '', '', '', '1', 'conv factor available', '8',
-             'using fct data from base term food code'],
-            ['8007_1', '225281cf-ae7e-4d20-9516-0e64c228467e', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'female', '1', '14', '6-59 months', 'no', 'yes', 'yes', 'rural', '2013002', 'Khoai sọ,nấu canh', '',
-             '9bfdd247-a9a4-48bd-91a1-3ae83c164a32', '', '201399', 'food_item', 'yes', 'fourth_pass', '04:38',
-             '3 am - 5:59 am', '2', 'Roots, tubers, plantains and their products (2)', '', '', '', '', '', '', '',
-             '',
-             '', '', '', 'khoai sọ', 'Khoai sọ', 'nấu canh', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW',
-             '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', '40', 'Photos (g)', '1', 'NC06-01', '1', 'grams', '5', '', '', '', '', '', '', '',
-             '',
-             '24.6', 'base_term_food_code', 'yes', 'yes', 'no', 'food_code', '123', '121.33', '149.2359', '67.6',
-             '83.148', '33.8', '41.574', '2', 'using conversion factor from base term food code', '1',
-             'fct data available'],
-            ['8007_1', '225281cf-ae7e-4d20-9516-0e64c228467e', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'female', '1', '14', '6-59 months', 'no', 'yes', 'yes', 'rural', '', 'Bánh quy,Khác', '',
-             '075922bb-3837-4b56-88d4-f7f61478e891', '', '1200199', 'food_item', 'yes', 'fourth_pass', '09:49',
-             '9 am - 11:59 am', '', '', '', '', '', '', '', '', '', '', '', '', '', 'bánh quy', 'Bánh quy', 'Khác',
-             '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '40', 'Photos (g)', '1',
-             'NC06-01', '1', 'grams', '2', '', '', '', '', '', '', '', '', '5', 'base_term_food_code', 'no', 'yes',
-             'no', 'base_term_food_code', '10', '378.1', '37.81', '65.8', '6.58', '17.4', '1.74', '2',
-             'using conversion factor from base term food code', '2', 'using fct data from base term food code'],
-            ['8007_1', '225281cf-ae7e-4d20-9516-0e64c228467e', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'female', '1', '14', '6-59 months', 'no', 'yes', 'yes', 'rural', '', 'Bánh mỳ,Khác,không ngọt', '',
-             '522d98f5-7d58-49dd-96ec-d0ebf6821c06', '', '101299', 'food_item', 'yes', 'fourth_pass', '02:22',
-             '12 am - 2:59 am', '', '', '', '', '', '', '', '', '', '', '', '', '', 'bánh mì sandwich', 'Bánh mỳ',
-             'Khác', '', 'không ngọt', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW',
-             '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '40', 'Photos (g)', '5',
-             'NC08-05', '1', 'grams', '2', '', '', '', '', '', '', '', '', '24', 'base_term_food_code', 'no', 'no',
-             'no', 'none', '48', '', '', '', '', '', '', '2', 'using conversion factor from base term food code',
-             '8',
-             'using fct data from base term food code'],
-            ['444567_1', 'e22a2245-8370-45b5-aea7-f2b8728dbc5e', '2019-08-22', 'john', 'john', '2019-08-17',
-             'Complete',
-             'male', '22', '264', '15-49 years', 'no', 'no', 'no', 'urban', '5001', 'Bưởi', '',
-             '7bc8b6bc-cc6d-46a2-b552-ca068c3a31d3', '', '5001', 'food_item', 'yes', 'fourth_pass', '02:29',
-             '12 am - 2:59 am', '10', 'Fruits and their products (10)', '', '', '', '', '', '', '', '', '', '', '',
-             'bưởi', 'Bưởi', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW',
-             '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             '40',
-             'Photos (g)', '4', 'NC07-03', '1', 'grams', '3', '', '', '', '', '', '', '', '', '', 'none', 'yes',
-             'yes',
-             'no', 'food_code', '', '32.79999999999994', '', '34.2', '', '3.8', '', '8',
-             'no conversion factor available', '1', 'Success'],
-            ['555_1', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b13', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '2013002130250', 'Canh khoai sọ',
-             'Canh khoai sọ', '1b3033c5-9ccc-4c92-aa59-13f0d9bb6500', '1b3033c5-9ccc-4c92-aa59-13f0d9bb6500',
-             '2013002130250', 'std_recipe', 'no', 'fourth_pass', '18:19', '6 pm - 8:59 pm', '', '', '', '', '', '',
-             '',
-             '', '', '1b3033c5-9ccc-4c92-aa59-13f0d9bb6500', '2013002130250', '', '', 'canh khoai sọ',
-             'I DONT KNOW',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '50',
-             'Proxy weight (g) - water', '', '', '200', 'grams', '1', '', '', '', '', '', '', '5', '1', '1',
-             'food_code', 'no', 'no', 'no', 'none', '200', '', '73.97842', '', '84.71156', '', '102.6692', '1',
-             'conv factor available', '', ''],
-            ['555_1', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b13', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '2013', 'Khoai sọ,sống', '', '',
-             '1b3033c5-9ccc-4c92-aa59-13f0d9bb6500', '201399', 'food_item', 'yes', 'fourth_pass', '18:19',
-             '6 pm - 8:59 pm', '2', 'Roots, tubers, plantains and their products (2)', '', '', '', '', '', '',
-             'yes',
-             '1b3033c5-9ccc-4c92-aa59-13f0d9bb6500', '2013002130250', '0.3069', '200', '#N/A', 'Khoai sọ', 'sống',
-             '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '', '', '', '', '', '',
-             '', '',
-             '', '', '', '', '', '', '', '', '', 'yes', 'yes', 'no', 'food_code', '61.38', '118.9', '72.98082',
-             '37.2',
-             '22.83336', '12', '7.3656', '1', 'conv factor available', '1', 'fct data available'],
-            ['555_1', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b13', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '4082', 'Rau mùi tàu,sống', '', '',
-             '1b3033c5-9ccc-4c92-aa59-13f0d9bb6500', '408299', 'food_item', 'yes', 'fourth_pass', '18:19',
-             '6 pm - 8:59 pm', '9', 'Vegetables and their products (9)', '', '', '', '', '', '', 'yes',
-             '1b3033c5-9ccc-4c92-aa59-13f0d9bb6500', '2013002130250', '0.003', '200', '#N/A', 'Rau mùi tàu',
-             'sống', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '', '', '', '', '', '',
-             '', '',
-             '', '', '', '', '', '', '', '', '', 'yes', 'no', 'no', 'food_code', '0.6', '27.60000000000004',
-             '0.1656',
-             '4.9', '0.0294', '64.3', '0.3858', '1', 'conv factor available', '1', 'fct data available'],
-            ['555_1', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b13', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '4038', 'Hành,lá (hành hoa),sống',
-             '', '',
-             '1b3033c5-9ccc-4c92-aa59-13f0d9bb6500', '403799', 'food_item', 'yes', 'fourth_pass', '18:19',
-             '6 pm - 8:59 pm', '9', 'Vegetables and their products (9)', '', '', '', '', '', '', 'yes',
-             '1b3033c5-9ccc-4c92-aa59-13f0d9bb6500', '2013002130250', '0.016', '200', '#N/A', 'Hành',
-             'lá (hành hoa)',
-             '', 'sống', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW',
-             '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '', '', '', '', '', '', '', '', '', '',
-             '',
-             '', '', '', '', '', '', 'yes', 'yes', 'no', 'food_code', '3.2', '26.00000000000001', '0.832', '13.9',
-             '0.4448', '68.2', '2.1824', '1', 'conv factor available', '1', 'fct data available'],
-            ['555_1', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b13', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '13025', 'Gia vị mặn,Bột canh', '',
-             '',
-             '1b3033c5-9ccc-4c92-aa59-13f0d9bb6500', '1302599', 'food_item', 'yes', 'fourth_pass', '18:19',
-             '6 pm - 8:59 pm', '13', 'Spices and condiments (13)', '', '', '', '', '', '', 'yes',
-             '1b3033c5-9ccc-4c92-aa59-13f0d9bb6500', '2013002130250', '0.003', '200', '#N/A', 'Gia vị mặn',
-             'Bột canh',
-             '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '', '', '', '', '', '',
-             '', '',
-             '', '', '', '', '', '', '', '', '', 'yes', 'yes', 'no', 'food_code', '0.6', '0', '0', '32.5', '0.195',
-             '40.3', '0.2418', '1', 'conv factor available', '1', 'fct data available'],
-            ['555_1', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b13', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '14018', 'Nước', '', '',
-             '1b3033c5-9ccc-4c92-aa59-13f0d9bb6500', '14018', 'food_item', 'yes', 'fourth_pass', '18:19',
-             '6 pm - 8:59 pm', '14', 'Beverages (14)', '', '', '', '', '', '', 'yes',
-             '1b3033c5-9ccc-4c92-aa59-13f0d9bb6500', '2013002130250', '0.6801', '200', '#N/A', 'Nước',
-             'I DONT KNOW',
-             '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '', '', '', '', '', '',
-             '', '',
-             '', '', '', '', '', '', '', '', '', 'yes', 'no', 'no', 'food_code', '136.02', '0', '0', '45',
-             '61.209',
-             '68', '92.4936', '1', 'conv factor available', '1', 'fct data available'],
-            ['555_1', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b13', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '9001008060081', 'Trứng gà ta ốp la',
-             'Trứng gà ta ốp la', '0b77fb58-f265-4897-b817-3a077d81dc3f', '0b77fb58-f265-4897-b817-3a077d81dc3f',
-             '9001008060081', 'std_recipe', 'no', 'fourth_pass', '06:34', '6 am - 8:59 am', '', '', '', '', '', '',
-             '',
-             '', '', '0b77fb58-f265-4897-b817-3a077d81dc3f', '9001008060081', '', '', 'trứng ốp la', 'I DONT KNOW',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '10',
-             'Direct weight (g)', '', '', '68', 'grams', '1', '', '', '', '', '', '', '2', '1', '1', 'food_code',
-             'no',
-             'no', 'no', 'none', '68', '', '', '', '', '', '', '1', 'conv factor available', '', ''],
-            ['555_1', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b13', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '6023', 'Dầu,Hỗn hợp', '', '',
-             '0b77fb58-f265-4897-b817-3a077d81dc3f', '602399', 'food_item', 'yes', 'fourth_pass', '06:34',
-             '6 am - 8:59 am', '11', 'Fats and oils (11)', '', '', '', '', '', '', 'yes',
-             '0b77fb58-f265-4897-b817-3a077d81dc3f', '9001008060081', '0.1915', '68', '#N/A', 'Dầu', 'Hỗn hợp', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '', '', '', '', '', '',
-             '', '',
-             '', '', '', '', '', '', '', '', '', 'yes', 'yes', 'no', 'food_code', '13.022', '900', '117.198',
-             '21.9',
-             '2.851818', '47', '6.12034', '1', 'conv factor available', '1', 'fct data available'],
-            ['555_1', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b13', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '9001', 'Egg,local hen,raw', '', '',
-             '0b77fb58-f265-4897-b817-3a077d81dc3f', '900188', 'food_item', 'yes', 'fourth_pass', '06:34',
-             '6 am - 8:59 am', '', '', '', '', '', '', '', '', 'yes', '0b77fb58-f265-4897-b817-3a077d81dc3f',
-             '9001008060081', '1.0426', '68', '#N/A', 'Trứng', 'gà ta', '', 'sống', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '',
-             '',
-             'no', 'no', 'no', 'food_code', '70.8968', '', '', '', '', '', '', '1', 'conv factor available', '8',
-             'using fct data from base term food code'],
-            ['1234_1', '738d5457-3517-4ef0-b21e-65a7ded2e8d1', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'male', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '', 'ĐẬU PHỤ RÁN XÀO SẢ ỚT',
-             'ĐẬU PHỤ RÁN XÀO SẢ ỚT', 'cb9c5609-379e-47bb-9cec-349171a4fc20',
-             'cb9c5609-379e-47bb-9cec-349171a4fc20',
-             '', 'non_std_recipe', 'no', 'fourth_pass', '21:13', '9 pm - 8:59 pm', '', '', '', '', '', '', '', '',
-             '',
-             'cb9c5609-379e-47bb-9cec-349171a4fc20', '', '', '', 'đậu phụ rán xào sả ớt', '', '', '', '', '', '',
-             '',
-             '', '', '', '', '', '', '', '', '', '', '', '', '', '', '51', 'Proxy weight (g) - rice', '', '', '36',
-             'grams', '1', '51', 'Proxy weight (g) - rice', '', '', '200', '0.18', '2', '', '', 'none', 'no', 'no',
-             'no', 'none', '5.939406', '', '4.29617034', '', '3.756404322', '', '1.098970092', '8',
-             'no conversion factor available', '7', 'ingredient(s) contain fct data gaps'],
-            ['1234_1', '738d5457-3517-4ef0-b21e-65a7ded2e8d1', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'male', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '13039', '(i) Gia vị,Ớt (tươi)', '',
-             '005b9946-abe4-411d-bb8f-5dbe3217523c', 'cb9c5609-379e-47bb-9cec-349171a4fc20', '1300199',
-             'food_item',
-             'yes', 'fourth_pass', '21:13', '9 pm - 8:59 pm', '', '', '', '', '', '', '', '', 'yes',
-             'cb9c5609-379e-47bb-9cec-349171a4fc20', '', '', '', '(i) ớt', 'Gia vị', 'Ớt (tươi)', '',
-             'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '40', 'Photos (g)', '1', 'NC06-01', '1',
-             'grams',
-             '1', '', '', '', '', '', '', '0', '1.9998', '', 'food_code', 'no', 'yes', 'no', 'base_term_food_code',
-             '0.359964', '0', '0', '68.6', '0.246935304', '75.9', '0.273212676', '1', 'conv factor available', '2',
-             'using fct data from base term food code'],
-            ['1234_1', '738d5457-3517-4ef0-b21e-65a7ded2e8d1', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'male', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '13024', '(i) Gia vị,Sả', '',
-             '190bf2a1-169d-4344-87d5-57dd3518edc2', 'cb9c5609-379e-47bb-9cec-349171a4fc20', '1300199',
-             'food_item',
-             'yes', 'fourth_pass', '21:13', '9 pm - 8:59 pm', '13', 'Spices and condiments (13)', '', '', '', '',
-             '',
-             '', 'yes', 'cb9c5609-379e-47bb-9cec-349171a4fc20', '', '', '', '(i) xả', 'Gia vị', 'Sả', '',
-             'I DONT KNOW',
-             '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '40', 'Photos (g)', '1', 'NC06-01', '1',
-             'grams',
-             '1', '', '', '', '', '', '', '0', '30.9969', '', 'food_code', 'yes', 'yes', 'no', 'food_code',
-             '5.579442',
-             '77', '4.29617034', '62.9', '3.509469018', '14.8', '0.825757416', '1', 'conv factor available', '1',
-             'fct data available'],
-            ['8007_1', '225281cf-ae7e-4d20-9516-0e64c228467e', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'female', '1', '14', '6-59 months', 'no', 'yes', 'yes', 'rural', '', 'QUẢ SUNG MUỐI', 'QUẢ SUNG MUỐI',
-             '01e36d8e-df92-41e3-a882-250af3c096d8', '01e36d8e-df92-41e3-a882-250af3c096d8', '', 'non_std_recipe',
-             'no',
-             'fourth_pass', '06:30', '6 am - 8:59 am', '', '', '', '', '', '', '', '', '',
-             '01e36d8e-df92-41e3-a882-250af3c096d8', '', '', '', 'quả sung muối', '', '', '', '', '', '', '', '',
-             '',
-             '', '', '', '', '', '', '', '', '', '', '', '', '51', 'Proxy weight (g) - rice', '', '', '32',
-             'grams',
-             '1', '51', 'Proxy weight (g) - rice', '', '', '110', '0.2909090909', '5', '', '', 'none', 'no', 'no',
-             'no',
-             'none', '', '', '', '', '', '', '', '8', 'no conversion factor available', '', ''],
-            ['8007_1', '225281cf-ae7e-4d20-9516-0e64c228467e', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'female', '1', '14', '6-59 months', 'no', 'yes', 'yes', 'rural', '', '(I) QUẢ SUNY,CẢ VỎ', '',
-             '04b0d499-cbfb-47a0-bb59-299231d6dbc8', '01e36d8e-df92-41e3-a882-250af3c096d8', '',
-             'non_std_food_item',
-             'yes', 'fourth_pass', '06:30', '6 am - 8:59 am', '', '', '', '', '', '', '', '', 'yes',
-             '01e36d8e-df92-41e3-a882-250af3c096d8', '', '', '', '(i) quả suny', '', '', '', '', '', '', '', '',
-             '', '',
-             '', '', '', '', '', '', '', '', '', '', '', '51', 'Proxy weight (g) - rice', '', '', '113', 'grams',
-             '1',
-             '', '', '', '', '', '', '0', '', '', 'none', 'no', 'no', 'no', 'none', '', '', '', '', '', '', '',
-             '8',
-             'no conversion factor available', '8', 'using fct data from base term food code'],
-            ['8007_1', '225281cf-ae7e-4d20-9516-0e64c228467e', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'female', '1', '14', '6-59 months', 'no', 'yes', 'yes', 'rural', '13017', '(i) Mắm,Nước mắm cá', '',
-             'c8d77ab1-5ae0-46ff-ab5a-2843347ffd0f', '01e36d8e-df92-41e3-a882-250af3c096d8', '1301199',
-             'food_item',
-             'yes', 'fourth_pass', '06:30', '6 am - 8:59 am', '13', 'Spices and condiments (13)', '', '', '', '',
-             '',
-             '', 'yes', '01e36d8e-df92-41e3-a882-250af3c096d8', '', '', '', '(i) nước mắm', 'Mắm', 'Nước mắm cá',
-             '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '50',
-             'Proxy weight (g) - water', '', '', '26', 'grams', '1', '', '', '', '', '', '', '0', '1', '1',
-             'food_code',
-             'yes', 'yes', 'no', 'food_code', '7.5636363636', '34.89', '2.6389527273', '51.2', '3.8725818182',
-             '18.5',
-             '1.3992727273', '1', 'conv factor available', '1', 'fct data available'],
-            ['8007_1', '225281cf-ae7e-4d20-9516-0e64c228467e', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'female', '1', '14', '6-59 months', 'no', 'yes', 'yes', 'rural', '13034', '(i) Giấm (gạo/bỗng/táo….)',
-             '',
-             '6929991b-2a7a-4fc2-b5bf-c74462643000', '01e36d8e-df92-41e3-a882-250af3c096d8', '13034', 'food_item',
-             'yes', 'fourth_pass', '06:30', '6 am - 8:59 am', '13', 'Spices and condiments (13)', '', '', '', '',
-             '',
-             '', 'yes', '01e36d8e-df92-41e3-a882-250af3c096d8', '', '', '', '(i) giấm', 'Giấm (gạo/bỗng/táo….)',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '50',
-             'Proxy weight (g) - water', '', '', '10', 'grams', '1', '', '', '', '', '', '', '0', '1.01', '1.01',
-             'food_code', 'yes', 'yes', 'no', 'food_code', '2.9381818182', '0', '0', '36', '1.0577454545', '17',
-             '0.4994909091', '1', 'conv factor available', '1', 'fct data available'],
-            ['8007_1', '225281cf-ae7e-4d20-9516-0e64c228467e', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'female', '1', '14', '6-59 months', 'no', 'yes', 'yes', 'rural', '12014', '(i) Đường,kính', '',
-             'cc39cbf9-4d38-4764-93ee-a6a30b326fea', '01e36d8e-df92-41e3-a882-250af3c096d8', '1201499',
-             'food_item',
-             'yes', 'fourth_pass', '06:30', '6 am - 8:59 am', '12', 'Sweets and sugars (12)', '', '', '', '', '',
-             '',
-             'yes', '01e36d8e-df92-41e3-a882-250af3c096d8', '', '', '', '(i) đường', 'Đường', 'kính', '',
-             'I DONT KNOW',
-             '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '10', 'Direct weight (g)', '', '', '34',
-             'grams',
-             '1', '', '', '', '', '', '', '0', '1', '1', 'food_code', 'yes', 'yes', 'no', 'food_code',
-             '9.8909090909',
-             '397.2', '39.2866909091', '26.1', '2.5815272727', '17.8', '1.7605818182', '1',
-             'conv factor available',
-             '1', 'fct data available'],
-            ['8007_1', '225281cf-ae7e-4d20-9516-0e64c228467e', '2019-08-22', 'sue', 'sue', '2019-08-17',
-             'Complete',
-             'female', '1', '14', '6-59 months', 'no', 'yes', 'yes', 'rural', '', '(i) Bột,Khác', '',
-             '88fe71d9-fbde-4962-b34c-2ff6918285d3', '01e36d8e-df92-41e3-a882-250af3c096d8', '101799', 'food_item',
-             'yes', 'fourth_pass', '06:30', '6 am - 8:59 am', '', '', '', '', '', '', '', '', 'yes',
-             '01e36d8e-df92-41e3-a882-250af3c096d8', '', '', '', '(i) bột canh', 'Bột', 'Khác', '', 'I DONT KNOW',
-             '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '10', 'Direct weight (g)', '', '', '10',
-             'grams',
-             '1', '', '', '', '', '', '', '0', '', '1', 'base_term_food_code', 'no', 'yes', 'no',
-             'base_term_food_code',
-             '2.9090909091', '360.4', '10.4843636364', '48.9', '1.4225454545', '16.6', '0.4829090909', '2',
-             'using conversion factor from base term food code', '2', 'using fct data from base term food code'],
-            ['555_2', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b14', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '', 'NƯỚC NGHỆ MẬT ONG',
-             'NƯỚC NGHỆ MẬT ONG', '36ce2096-0bda-4c15-a47f-19dc955e76da', '36ce2096-0bda-4c15-a47f-19dc955e76da',
-             '',
-             'non_std_recipe', 'no', 'fourth_pass', '09:24', '9 am - 11:59 am', '', '', '', '', '', '', '', '', '',
-             '36ce2096-0bda-4c15-a47f-19dc955e76da', '', '', '', 'nước nghệ mật ong', '', '', '', '', '', '', '',
-             '',
-             '', '', '', '', '', '', '', '', '', '', '', '', '', '50', 'Proxy weight (g) - water', '', '', '190',
-             'grams', '1', '10', 'Direct weight (g)', '', '', '190', '', '2', '', '', 'none', 'no', 'no', 'no',
-             'none',
-             '', '', '', '', '', '', '', '8', 'no conversion factor available', '8',
-             'using fct data from base term food code'],
-            ['555_2', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b14', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '', '(I) BỘT NGHỆ,BỘT NGHỆ XAY', '',
-             '9c440036-eb8b-461a-94d8-1afdbebe4e9c', '36ce2096-0bda-4c15-a47f-19dc955e76da', '',
-             'non_std_food_item',
-             'yes', 'fourth_pass', '09:24', '9 am - 11:59 am', '', '', '', '', '', '', '', '', 'yes',
-             '36ce2096-0bda-4c15-a47f-19dc955e76da', '', '', '', '(i) bột nghệ', '', '', '', '', '', '', '', '',
-             '', '',
-             '', '', '', '', '', '', '', '', '', '', '', '10', 'Direct weight (g)', '', '', '6', 'grams', '1', '',
-             '',
-             '', '', '', '', '0', '', '', 'none', 'no', 'no', 'no', 'none', '', '', '', '', '', '', '', '8',
-             'no conversion factor available', '8', 'using fct data from base term food code'],
-            ['555_2', 'f3bfc88d-6cb1-4ae8-acdc-9715a9411b14', '2019-08-22', 'mary', 'mary', '2019-08-12',
-             'Complete',
-             'female', '23', '276', '15-49 years', 'no', 'no', 'no', 'urban', '12026', '(i) Mật ong', '',
-             'a593172b-2c0c-45d5-9533-6ac8cba0b14d', '36ce2096-0bda-4c15-a47f-19dc955e76da', '12026', 'food_item',
-             'yes', 'fourth_pass', '09:24', '9 am - 11:59 am', '12', 'Sweets and sugars (12)', '', '', '', '', '',
-             '',
-             'yes', '36ce2096-0bda-4c15-a47f-19dc955e76da', '', '', '', '(i) mật ong', 'Mật ong', 'I DONT KNOW',
-             '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '',
-             'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', 'I DONT KNOW', '', '50',
-             'Proxy weight (g) - water', '', '', '8', 'grams', '1', '', '', '', '', '', '', '0', '1.43', '1.43',
-             'food_code', 'yes', 'yes', 'no', 'food_code', '', '326.8', '', '73.6', '', '49.4', '', '1',
-             'conv factor available', '1', 'fct data available'],
-            ['12345_1', 'fbcc5cd9-3d8c-43c8-908d-abb1c0b21089', '2019-08-22', 'mary', 'mary', '2019-08-17',
-             'Complete',
-             'male', '23', '276', '15-49 years', 'yes', 'no', 'no', 'urban', '',
-             'NƯỚC ĐẬU,NƯỚC ĐẬU TƯƠI THÊM ĐƯỜNG VÀ ĐÁ', '', '2b4bb306-5fb5-48d8-8c76-8f6be6ae8a9a', '', '',
-             'non_std_food_item', 'no', 'fourth_pass', '06:17', '6 am - 8:59 am', '', '', '', '', '', '', '', '',
-             '',
-             '', '', '', '', 'nước đậu', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '',
-             '', '',
-             '', '50', 'Proxy weight (g) - water', '', '', '172', 'grams', '1', '', '', '', '', '', '', '0', '',
-             '',
-             'none', 'no', 'no', 'no', 'none', '', '', '', '', '', '', '', '8', 'no conversion factor available',
-             '8',
-             'using fct data from base term food code'],
-            ['222_1', 'fbcc5cd9-3d8c-43c8-908d-abb1c0b21077', '2019-08-22', 'mary', 'mary', '2019-08-17',
-             'Complete',
-             'male', '5', '61', '05-06 years', 'yes', 'no', 'no', 'urban', '',
-             'NƯỚC ĐẬU,NƯỚC ĐẬU TƯƠI THÊM ĐƯỜNG VÀ ĐÁ', '', '2b4bb306-5fb5-48d8-8c76-8f6be6ae8a9a', '', '',
-             'non_std_food_item', 'no', 'fourth_pass', '06:17', '6 am - 8:59 am', '', '', '', '', '', '', '', '',
-             '',
-             '', '', '', '', 'nước đậu', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '',
-             '', '',
-             '', '50', 'Proxy weight (g) - water', '', '', '172', 'grams', '1', '', '', '', '', '', '', '0', '',
-             '',
-             'none', 'no', 'no', 'no', 'none', '', '', '', '', '', '', '', '8', 'no conversion factor available',
-             '8',
-             'using fct data from base term food code'],
-            ['223_1', 'fbct5cd9-3d8c-43c8-908d-abb1c0b21078', '2019-08-22', 'mary', 'mary', '2019-08-17',
-             'Complete',
-             'male', '8', '97', '07 -10 years', 'yes', 'no', 'no', 'urban', '',
-             'NƯỚC ĐẬU,NƯỚC ĐẬU TƯƠI THÊM ĐƯỜNG VÀ ĐÁ', '', '2b4bb306-5fb5-48d8-8c76-8f6be6ae8a9a', '', '',
-             'non_std_food_item', 'no', 'fourth_pass', '06:17', '6 am - 8:59 am', '', '', '', '', '', '', '', '',
-             '',
-             '', '', '', '', 'nước đậu', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '',
-             '', '',
-             '', '50', 'Proxy weight (g) - water', '', '', '172', 'grams', '1', '', '', '', '', '', '', '0', '',
-             '',
-             'none', 'no', 'no', 'no', 'none', '', '', '', '', '', '', '', '8', 'no conversion factor available',
-             '8',
-             'using fct data from base term food code'],
-            ['224_1', 'fbcc5cd9-3d8c-43c8-908d-abb1c0b21079', '2019-08-22', 'mary', 'mary', '2019-08-17',
-             'Complete',
-             'male', '12', '145', '11-14 years', 'yes', 'no', 'no', 'urban', '',
-             'NƯỚC ĐẬU,NƯỚC ĐẬU TƯƠI THÊM ĐƯỜNG VÀ ĐÁ', '', '2b4bb306-5fb5-48d8-8c76-8f6be6ae8a9a', '', '',
-             'non_std_food_item', 'no', 'fourth_pass', '06:17', '6 am - 8:59 am', '', '', '', '', '', '', '', '',
-             '',
-             '', '', '', '', 'nước đậu', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '',
-             '', '',
-             '', '50', 'Proxy weight (g) - water', '', '', '172', 'grams', '1', '', '', '', '', '', '', '0', '',
-             '',
-             'none', 'no', 'no', 'no', 'none', '', '', '', '', '', '', '', '8', 'no conversion factor available',
-             '8',
-             'using fct data from base term food code'],
-            ['225_1', 'fbcc5cd9-3d8c-43c8-908d-abb1c0b21080', '2019-08-22', 'mary', 'mary', '2019-08-17',
-             'Complete',
-             'male', '0', '3', '0-5.9 months', 'yes', 'no', 'no', 'urban', '',
-             'NƯỚC ĐẬU,NƯỚC ĐẬU TƯƠI THÊM ĐƯỜNG VÀ ĐÁ', '', '2b4bb306-5fb5-48d8-8c76-8f6be6ae8a9a', '', '',
-             'non_std_food_item', 'no', 'fourth_pass', '06:17', '6 am - 8:59 am', '', '', '', '', '', '', '', '',
-             '',
-             '', '', '', '', 'nước đậu', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '',
-             '', '',
-             '', '50', 'Proxy weight (g) - water', '', '', '172', 'grams', '1', '', '', '', '', '', '', '0', '',
-             '',
-             'none', 'no', 'no', 'no', 'none', '', '', '', '', '', '', '', '8', 'no conversion factor available',
-             '8',
-             'using fct data from base term food code'],
-        ]
-
-
 class ConvFactorGapsSummaryData(GapsReportSummaryDataMixin):
     title = 'Conv Factor Gaps Summary by Food Type'
     slug = 'conv_factor_gaps_summary_by_food_type'
